Updates to seeds

+++ db/migrate/20230618071004_create_products.rb
@@ -1,7 +1,7 @@
 class CreateProducts < ActiveRecord::Migration[7.0]
   def change
     create_table :products do |t|
-      t.decimal :weight
+      t.float :weight
       t.string :sku, null: false, index: { unique: true }
       t.string :unit
       t.references :category, null: false, foreign_key: true
